Add files via upload
This python file simulates the orbits of the planetary bodies and the sun
using newtonian 2-body equations.  This file depends on python 2.7, vpython 
classic, scipy.integrate, scipy.constants, pylab, and the future version of
division.  If you try to run this file without any of these, the simulation
probably won't run correctly.

This project was initially completed for a college astrophysics course.  The 
cohesion and readability of the project reflected my poor programming skills.
This can be seen in solar_old.py, which is the original version.  This new
version reorganizes the project for readability and simplicity.  I've removed
copious redundancy through the implementation of a basic Class structure.

In the future, I plan on reducing run time so that
more datapoints can be used.